+++ package.json
@@ -24,7 +24,9 @@
   "homepage": "https://github.com/danmarshall/deckgl-typings#readme",
   "devDependencies": {
     "math.gl": "*",
+    "@luma.gl/core": "*",
     "@luma.gl/webgl": "*",
+    "@luma.gl/webgl-state-tracker": "*",
     "@types/node": "^12.12.5",
     "dts-generator": "^3.0.0",
     "glob": "^7.1.5",
