+++ test/index.d.ts
@@ -1,2 +1,4 @@
 /// <reference path="./math.gl.d.ts" />
 /// <reference path="./luma.gl/webgl.d.ts" />
+/// <reference path="./luma.gl/webgl-state-tracker.d.ts" />
+/// <reference path="./luma.gl/core.d.ts" />
