+++ test/luma.gl/core.d.ts
@@ -1,0 +1,5 @@
+/// <reference path="../../luma.gl__core/index.d.ts" />
+
+import * as LumaGlCore from '@luma.gl/core';
+
+type ls = LumaGlCore.SphereGeometry;
