+++ test/luma.gl/webgl-state-tracker.d.ts
@@ -1,0 +1,3 @@
+/// <reference path="../../luma.gl__webgl-state-tracker/index.d.ts" />
+
+import { getModifiedParameters } from '@luma.gl/webgl-state-tracker';
